+++ qiita_reader_ios/QiitaReader/Data/Repository/Auth/AuthRepository.swift
@@ -11,24 +11,19 @@ import Foundation
 protocol AuthRepositoryProtocol {
     func authorize() -> AnyPublisher<Authorize, AuthError>
     func generateAccessToken(authorizedCode: String) -> AnyPublisher<AuthToken, APIError>
-    func setAccessToken(_ value: String) -> AnyPublisher<Void, Error>
-    func getAccessToken() -> String?
-    func deleteAccessToken() -> AnyPublisher<Void, Error>
 }
 
 class AuthRepository: AuthRepositoryProtocol {
     private let authClient: AuthClientProtocol
     private let apiClient: APIClientProtocol
-    private let secureStorageClient: SecureStorageClientProtocol
+
 
     init(
         authClient: AuthClientProtocol = AuthClient(),
-        apiClient: APIClientProtocol = APIClient(),
-        secureStorageClient: SecureStorageClientProtocol = SecureStorageClient()
+        apiClient: APIClientProtocol = APIClient()
     ) {
         self.authClient = authClient
         self.apiClient = apiClient
-        self.secureStorageClient = secureStorageClient
     }
 
     func authorize() -> AnyPublisher<Authorize, AuthError> {
@@ -46,32 +41,5 @@ class AuthRepository: AuthRepositoryProtocol {
         return apiClient
             .start(request)
             .eraseToAnyPublisher()
-    }
-
-    func setAccessToken(_ value: String) -> AnyPublisher<Void, Error> {
-        let key = Env.Qiita.accessTokenStorageKey
-        return Result {
-            try secureStorageClient.setData(key: key, value: value)
-        }
-        .publisher
-        .eraseToAnyPublisher()
-    }
-
-    func getAccessToken() -> String? {
-        let key = Env.Qiita.accessTokenStorageKey
-        do {
-            return try secureStorageClient.getStringData(key: key)
-        } catch {
-            return nil
-        }
-    }
-
-    func deleteAccessToken() -> AnyPublisher<Void, Error> {
-        let key = Env.Qiita.accessTokenStorageKey
-        return Result {
-            try secureStorageClient.deleteData(key: key)
-        }
-        .publisher
-        .eraseToAnyPublisher()
     }
 }
